Refactor to async/await
Change unpinning of messages to be conditioned on whether a message
was pinned and not if we are allowed to pin messages.

Return 'false' in methods that should not reply to the message because they
send messages to the channel instead.

Change name of rollers to use their nickname for the server, if one has been set,
otherwise fallback to their username.